refactor header-footer.tsx
