+++ apps/frontend/src/routes/{-$lang}/_layout/_authed/admin/-components/site-navigation-editor/types.ts
@@ -1,0 +1,14 @@
+import type { NavigationGroup, NavigationItem } from "@/config/site-navigation";
+import type { NavbarCommittedGroup } from "@/config/site-navigation-admin";
+
+export type NavbarGroupWithItems = NavbarCommittedGroup;
+
+export type FooterGroupWithItems = {
+  group: NavigationGroup;
+  items: Array<{
+    item: NavigationItem;
+    enabled: boolean;
+  }>;
+};
+
+export type ItemsRecord = Record<string, string[]>;
